move hook to correct directory layer
